Release 2.0.11 App sort fix

+++ android/src/com/volla/launcher/worker/AppWorker.java
@@ -112,7 +112,9 @@ public class AppWorker
                                         timeInForeground = timeInForeground + 10; // fall back for missing stats
                                     }
 
-                                    appInfo.put("statistic", timeInForeground);
+                                    //Log.d(TAG, "Statistic: " + timeInForeground);
+
+                                    appInfo.put("statistic", (int)timeInForeground);
                                     appList.add(appInfo);
                                 }
                             }
